SupplyChart as a React Component
